+++ src/hooks/CoffeeService.jsx
@@ -1,11 +1,10 @@
 // CoffeeService.jsx
-import CoffeeModel from './models/CoffeeModel';
+import { fetchAllWithReviews } from './CoffeeModelService.jsx';
 
-/**
- * Fetch coffees + their reviews via CoffeeModel
- */
-export async function fetchCoffees() {
-  const results = await CoffeeModel.fetchAllWithReviews();
+ //Fetch coffees + their reviews via fetchAllwithReviews
+
+ export async function fetchCoffees() {
+  const results = await fetchAllWithReviews();
   return results.map(c => ({
     id: c.id,
     name: c.get('name'),
@@ -18,3 +17,21 @@ export async function fetchCoffees() {
     })) || []
   }));
 }
+
+
+//Coffee and CoffeeReviews are 2 seperate classes. Through documentation, we tried accessing the
+// values through query.include('reviews'), which was stored as an array of pointers, like so - 
+// [
+//   {
+//     "__type": "Pointer",
+//     "className": "CoffReview",
+//     "objectId": "LfPZo7IIs5"
+//   },
+//   {
+//     "__type": "Pointer",
+//     "className": "CoffReview",
+//     "objectId": "fFRPJKzg8r"
+//   }
+// ]
+
+// and directly access the values in CoffReview class.